requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
autopep8==1.4
bcrypt==3.1.7
cachetools==4.0.0
certifi==2018.4.16
cffi==1.14.0
chardet==3.0.4
click==6.7
cryptography==2.8
Django==2.1.1
dlib==19.15.0
fabric==2.5.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-WTF==0.14.3
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
image==1.5.25
invoke==1.4.1
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
paramiko==2.7.1
Pillow==5.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
pytz==2018.5
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.14.1
wrapt==1.11.2
WTForms==2.2.1
